pyproject.toml
[tool.poetry]
name = "nlu"
version = "0.1.0"
description = "For amazon Teckathon 2022"
authors = ["Asseel Naji"]
license = "wtfpl"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
openai = "^0.20.0"
boto3 = "^1.24.38"
PyQt5 = "^5.15.7"
pyowm = "^3.3.0"
google-cloud-speech = "^2.15.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
